feat: modularize subcommands
