--- src/commands/command.ts
@@ -1,333 +1,0 @@
-import { ChatInputCommandInteraction, SlashCommandSubcommandBuilder, SlashCommandSubcommandGroupBuilder } from "discord.js"
-import ApplicationCommand from "../types/ApplicationCommand"
-import database from "../utils/database"
-import embeds from "../utils/embeds"
-import commands from "../utils/commands"
-import format from "../utils/format"
-import { stripIndent, stripIndents } from "common-tags"
-import { permissions } from "../config.json"
-import { client } from ".."
-const {
-	SlashCommandBuilder,
-	EmbedBuilder,
-} = require("discord.js")
-
-export default new ApplicationCommand({
-	settings: {
-		ownerOnly: true
-	},
-	data: new SlashCommandBuilder()
-		.setName("command")
-		.setDescription("Configure my commands")
-		.addSubcommandGroup((group: SlashCommandSubcommandGroupBuilder) =>
-			group
-				.setName("alias")
-				.setDescription("alias")
-				.addSubcommand((command: SlashCommandSubcommandBuilder) =>
-					command
-						.setName("create")
-						.setDescription("create an alias")
-						.addStringOption((option) =>
-							option
-								.setName("alias")
-								.setDescription("alias to create")
-								.setRequired(true),
-						)
-						.addStringOption((option) =>
-							option
-								.setName("commandname")
-								.setDescription("commandname")
-								.setRequired(true),
-						)
-						.addStringOption((option) =>
-							option.setName("group").setDescription("group"),
-						)
-						.addStringOption((option) =>
-							option
-								.setName("subcommand")
-								.setDescription("subcommand"),
-						)
-						.addStringOption((option) =>
-							option
-								.setName("defaultoptions")
-								.setDescription("defaultoptions"),
-						)
-						.addBooleanOption((option) =>
-							option
-								.setName("hidealloptions")
-								.setDescription("hide options?"),
-						)
-						.addStringOption((option) =>
-							option
-								.setName("description")
-								.setDescription("override description"),
-						),
-				)
-				.addSubcommand((command: SlashCommandSubcommandBuilder) =>
-					command
-						.setName("remove")
-						.setDescription("remove an alias")
-						.addStringOption((option) =>
-							option
-								.setName("alias")
-								.setDescription("alias to remove")
-								.setRequired(true),
-						),
-				)
-				.addSubcommand((command) =>
-					command.setName("list").setDescription("list all aliases"),
-				),
-		)
-		.addSubcommandGroup((group: SlashCommandSubcommandGroupBuilder) =>
-			group
-				.setName("command")
-				.setDescription("command")
-				.addSubcommand((command) =>
-					command
-						.setName("enable")
-						.setDescription("enable a disabled command")
-						.addStringOption((option) =>
-							option
-								.setName("command")
-								.setDescription("command to enable")
-								.setRequired(true),
-						),
-				)
-				.addSubcommand((command) =>
-					command
-						.setName("disable")
-						.setDescription("disable a command")
-						.addStringOption((option) =>
-							option
-								.setName("command")
-								.setDescription("command to disable")
-								.setRequired(true),
-						),
-				)
-				.addSubcommand((command) =>
-					command
-						.setName("refresh")
-						.setDescription("refresh guild commands")),
-		)
-		// .addSubcommand((command: SlashCommandSubcommandBuilder) =>
-		// 	command
-		// 		.setName("run")
-		// 		.setDescription("run a command")
-		// 		.addStringOption((option) =>
-		// 			option
-		// 				.setName("command")
-		// 				.setDescription("command to run")
-		// 				.setRequired(true),
-		// 		),
-		// )
-		.addSubcommand((command) =>
-			command.setName("list").setDescription("list all commands"),
-		)
-		.addSubcommand((command) =>
-			command
-				.setName("parse")
-				.setDescription("wip")
-				.addStringOption((option) =>
-					option
-						.setName("text")
-						.setDescription("text")
-						.setRequired(true),
-				),
-		),
-	async execute(interaction: ChatInputCommandInteraction): Promise<void> {
-		switch (interaction.options.getSubcommandGroup()) {
-			case "alias": {
-				const alias = interaction.options.getString("alias").toLowerCase()
-				console.log("alias: " + alias)
-
-				const guildID = interaction.guild.id
-				const path = `.guilds.${guildID}.commands.aliases`
-				const aliasPath = path + "." + alias
-				var value: string
-				var exists: boolean
-
-				if (alias) {
-					try {
-						value = await database.get(aliasPath)
-					} catch (e) {
-						value = undefined
-					}
-					console.log(`value: ${value}`)
-					if (value) exists = true
-					else exists = false
-				}
-
-
-
-				switch (interaction.options.getSubcommand()) {
-					case "create": {
-						if (exists)
-							throw new Error(`$${aliasPath} already exists`)
-
-						// check if there is a command already named as this alias, and prevent that
-						if (client.commands.has(alias))
-							throw new Error(`${alias} is an already existing command`)
-						if ((await interaction.guild.commands.fetch()).has(alias))
-							throw new Error(`${alias} is an already existing guild command`)
-
-						try {
-							const command =
-								interaction.options.getString("commandname")
-							const group = interaction.options.getString("group")
-							const subcommand =
-								interaction.options.getString("subcommand")
-							const defaultoptions = JSON.parse(
-								interaction.options.getString("defaultoptions"),
-							)
-							const hidealloptions =
-								interaction.options.getBoolean("hidealloptions")
-							const description =
-								interaction.options.getString("description")
-							console.log(defaultoptions)
-
-							const data = {
-								commandname: command,
-								group: group,
-								subcommand: subcommand,
-								defaultoptions: defaultoptions || [],
-								hidedefaults: true,
-								hidealloptions: hidealloptions,
-								description: description,
-							}
-							await database.set(aliasPath, data)
-							commands.refreshGuild(guildID)
-
-							interaction.reply({
-								embeds: embeds.successEmbed(
-									"Created alias successfully",
-								),
-							})
-						} catch (err) {
-							// interaction.reply({
-							// 	embeds: embeds.errorEmbed(
-							// 		"Creating alias",
-							// 		err,
-							// 	),
-							// })
-							throw new Error(
-								`Alias $${aliasPath} could not be created\n${err}`,
-							)
-						}
-						break
-					}
-					case "remove": {
-						if (!exists)
-							throw new Error(`${aliasPath} does not exist`)
-						try {
-							await database.del(aliasPath)
-							commands.refreshGuild(guildID)
-							interaction.reply({
-								embeds: embeds.successEmbed(
-									"Removed alias successfully",
-								),
-							})
-						} catch (err) {
-							interaction.reply({
-								embeds: embeds.errorEmbed(
-									"Removing alias",
-									err,
-								),
-							})
-						}
-
-						break
-					}
-					case "list": {
-						value = await database.get(path)
-						await interaction.reply({
-							embeds: embeds.messageEmbed(
-								"List:",
-								JSON.stringify(value),
-							),
-						})
-						break
-					}
-				}
-				break
-			}
-			case "command": {
-				switch (interaction.options.getSubcommand()) {
-					case "enable": {
-						break
-					}
-					case "disable": {
-						break
-					}
-					case "refresh": {
-						commands.refreshGuild(interaction.guild.id)
-						interaction.reply({
-							embeds: embeds.successEmbed(
-								"Refreshed guild commands",
-							),
-						})
-						break
-					}
-					default: {
-						throw new Error("it broke")
-					}
-				}
-				break
-			}
-
-			default: {
-				switch (interaction.options.getSubcommand()) {
-					case "run": {
-						break
-					}
-					case "list": {
-						const lines = []
-						const commandList = await commands.get()
-						for (
-							let i = 0, len = commandList.length;
-							i < len;
-							i++
-						) {
-							const command = commandList[i]
-							if (typeof command.execute === "function") {
-								console.log(command + typeof command.execute)
-								lines.push(
-									`${command.data.name} - "${command.data.description}"`,
-								)
-							} else {
-								console.log("not a slash command")
-							}
-						}
-						const message = lines.join("\n")
-						const messages = format.splitMessage(message)
-						for (let i = 0, len = messages.length; i < len; i++) {
-							if (i == 0) interaction.reply(messages[i])
-							else interaction.followUp(messages[i])
-						}
-						break
-					}
-					case "parse": {
-						if (interaction.user.id !== permissions.botowner) return
-
-						const text = interaction.options.getString("text")
-						const [commandName, group, subcommand, foundoptions] =
-							await commands.textToCommandParser(text)
-						console.log([
-							commandName,
-							group,
-							subcommand,
-							foundoptions,
-						])
-						commands.run(
-							interaction,
-							"slash",
-							commandName,
-							group,
-							subcommand,
-							foundoptions,
-						)
-					}
-				}
-			}
-		}
-	},
-})
